requirements.txt
streamlit
langchain
python-dotenv
pandas
langchain_community
pypdf
groq
langchain_groq
torch
transformers
chromadb
sentence_transformers
InstructorEmbedding
pdf2image
auto-gptq
poppler-utils
pytesseract
PyPDF2
streamlit
pypdf
pdf2image
pytesseract
pandas
langchain-groq
pillow
transformers
